+++ blocking_step.go
@@ -71,21 +71,7 @@ func waitForUnlockNotify(tls *libc.TLS, db uintptr, unPtr uintptr) ResultCode {
 	}
 	un.fired = false
 
-	// The following is a conversion from function value to uintptr. It assumes
-	// the memory representation described in https://golang.org/s/go11func.
-	//
-	// It does this by doing the following in order:
-	// 1) Create a Go struct containing a pointer to a pointer to
-	//    unlockNotifyCallback. It is assumed that the pointer to
-	//    unlockNotifyCallback will be stored in the read-only data section and
-	//    thus will not move.
-	// 2) Convert the pointer to the Go struct to a pointer to uintptr through
-	//    unsafe.Pointer. This is permitted via Rule #1 of unsafe.Pointer.
-	// 3) Dereference the pointer to uintptr to obtain the function value as a
-	//    uintptr. This is safe as long as function values are passed as pointers.
-	cbPtr := *(*uintptr)(unsafe.Pointer(&struct {
-		f func(*libc.TLS, uintptr, int32)
-	}{unlockNotifyCallback}))
+	cbPtr := cFuncPointer(unlockNotifyCallback)
 
 	res := ResultCode(lib.Xsqlite3_unlock_notify(tls, db, cbPtr, unPtr))