+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/zombiezen/go-sqlite/compare/v0.12.0...main
 
+## [Unreleased][]
+
+### Added
+
+- Support user-defined collating sequences
+  ([#21](https://github.com/zombiezen/go-sqlite/issues/21)).
+
 ## [0.12.0][] - 2023-02-08
 
 Version 0.12 added support for the [online backup API](https://www.sqlite.org/backup.html).